feat: add event handler and read model L1 and refactor infrastructure.ts
