+++ examples/airline/reservations/package.json
@@ -16,6 +16,8 @@
     "@aws-cdk/aws-events": "^1.105.0",
     "@aws-cdk/aws-dynamodb": "^1.105.0",
     "@aws-cdk/core": "^1.105.0",
+    "@aws-sdk/client-dynamodb": "^3.16.0",
+    "@aws-sdk/util-dynamodb": "^3.16.0",
     "aws-cdk": "^1.105.0",
     "operations": "^0.1.0",
     "scheduling": "^0.1.0",
