+++ packages/stochastic-aws-serverless/src/index.ts
@@ -1,5 +1,9 @@
+export * from "./aggregate-construct"
+export * from "./bounded-context-construct"
+export * from "./command-construct"
+export * from "./component-construct"
 export * from "./dependency-construct"
 export * from "./event-binding"
+export * from "./event-store-construct"
 export * from "./event-store"
-export * from "./infrastructure"
-export * from "./runtime"
+export * from "./policy-construct"
