+++ packages/stochastic/src/index.ts
@@ -1,10 +1,11 @@
-export * from "./aggregate";
-export * from "./command";
-export * from "./component";
-export * from "./dependencies";
-export * from "./bounded-context";
-export * from "./event";
-export * from "./policy";
-export * from "./query";
-export * from "./read-model";
-export * from "./shape";
+export * from "./aggregate"
+export * from "./command"
+export * from "./component"
+export * from "./dependencies"
+export * from "./bounded-context"
+export * from "./event"
+export * from "./event-handler"
+export * from "./policy"
+export * from "./query"
+export * from "./read-model"
+export * from "./shape"
